get_time: add cc file
