+++ pbbslib/BUILD
@@ -67,6 +67,7 @@ cc_library(
 cc_library(
   name = "get_time",
   hdrs = ["get_time.h"],
+  srcs = ["get_time.cc"],
 )
 
 cc_library(
