+++ pbbslib/get_time.h
@@ -1,9 +1,6 @@
 #pragma once
 
-#include <stdlib.h>
 #include <sys/time.h>
-#include <iomanip>
-#include <iostream>
 #include <string>
 
 struct timer {
@@ -13,71 +10,27 @@ struct timer {
   std::string name;
   struct timezone tzp;
 
-  timer(std::string name = "# PBBS time", bool _start = true)
-  : total_time(0.0), on(false), name(name), tzp({0,0}) {
-    if (_start) start();
-  }
+  timer(const std::string& name = "# PBBS time", bool _start = true);
 
-  double get_time() {
-    timeval now;
-    gettimeofday(&now, &tzp);
-    return ((double) now.tv_sec) + ((double) now.tv_usec)/1000000.;
-  }
+  double get_time();
 
-  void start () {
-    on = 1;
-    last_time = get_time();
-  }
+  void start();
 
-  double stop () {
-    on = 0;
-    double d = (get_time()-last_time);
-    total_time += d;
-    return d;
-  }
+  double stop();
 
-  void reset() {
-     total_time=0.0;
-     on=0;
-  }
+  void reset();
 
-  double get_total() {
-    if (on) return total_time + get_time() - last_time;
-    else return total_time;
-  }
+  double get_total();
 
-  double get_next() {
-    if (!on) return 0.0;
-    double t = get_time();
-    double td = t - last_time;
-    total_time += td;
-    last_time = t;
-    return td;
-  }
+  double get_next();
 
-  void report(double time, std::string str) {
-    std::ios::fmtflags cout_settings = std::cout.flags();
-    std::cout.precision(4);
-    std::cout << std::fixed;
-    std::cout << name << ": ";
-    if (str.length() > 0)
-      std::cout << str << ": ";
-    std::cout << time << std::endl;
-    std::cout.flags(cout_settings);
-  }
+  void report(double time, const std::string& str);
 
-  void total() {
-    report(get_total(),"total");
-    total_time = 0.0;
-  }
+  void total();
 
-  void reportTotal(std::string str) {
-    report(get_total(), str);
-  }
+  void reportTotal(const std::string& str);
 
-  void next(std::string str) {
-    if (on) report(get_next(), str);
-  }
+  void next(const std::string& str);
 };
 
 static timer _tm;
